Move solver to code generated class
When independent states change after toggling switches,
the states need to be updated from outputs e.g. inductor
that has a switch opened in series needs to have its state
current to be set to zero. Previously the derivative was set
to 0 and the current was not visible in output but it
affected the next time the switch was closed

- Generate the state space into a class so that it is namespaced
automatically
- Add semicolon for netlist measured components for
easier parsing e.g. current dependent current source
the source current is not considered as measurement