Add option to store the circuit in json
With large circuits the generated C++ files
become so large that they can take several
minutes to compile. Add option to store the
symbolic circuit as json and load it at runtime,
change the symbolic values to numeric and
then calculate the state space matrices. The
json file is embedded to the executable with
a resource file so that it does not need to be
copied around. A .rc file is made for that but
user needs to give the resource id as there is
no way to know which ones are used if there
are multiple circuits in one exe. Maybe strings
can be used as ids? I dunno.

There are 2 risky places right now in the json
version. If a component name is substring of
another component name, the name may
be replaced wrong. Another is evaluateExpression
may not be correct for all inputs, have not tested
enough.