Try to speed variable timestepping with better matrix caching
If implicit integration step size limit is set, different step sizes
are integrated with Tustin to closest multiple and
the remaining with RKF e.g. 11.1us with 1 us is integrated in steps of 11 Tustin and 0.1 RKF. This allows caching the inverse matrices
that are multiples of the minimum step size and RKF only has
to solve small steps.

For caching the solver does not anymore check jacobian and
it is up to the user to update it when it changes. This way the
inverse coeff calculation only needs to check if dt has changed.