Add support for generating matrices dynamically
With a lot of switches in a model it simply is not anymore
possible to solve all possible combinations beforehand even
if switches are marked mutually exclusive. Add support for
code generation with json that only the case with all
switches opened is solved initially and when the generated
code is ran, it will call the generation script again for
combinations for which the matrices are missing and store
them in the json.
This way only the actually used switch combinations are
generated.