Allow some invalid circuits but print a warning
In a three-phase circuit with three switches and three current sources in series
as coupling elements, the circuit is not valid if all the switches are open. However,
that combination may not be interesting and making the circuit valid would require
adding snubbers that are not needed in normal use cases. For this purpose allow
the invalid circuit so that when the switches are closed, the circuit works without
adding extra snubbers.